requirements.txt
scikit-learn==0.21.3
miditoolkit==0.1.14
fastBPE
dtw-python
nltk
keras
tqdm
matplotlib
pretty_midi
PyYAML
pypianoroll
fairseq==0.10.2
torch==1.7.1
transformers==3.5.1
pypinyin==0.39.1
jieba==0.42.1
tensorboard==2.4.0
scipy==1.3.1
librosa
pyworld
soundfile
pyOpenSSL
secrets
